--- lib/Pages/Midleware.dart
@@ -1,32 +1,0 @@
-import 'package:flutter/material.dart';
-import 'package:flutterboiler/Configs/Colors.dart';
-import 'package:flutterboiler/Utils/PrintUtils.dart';
-
-class MiddleWare extends StatefulWidget {
-  MiddleWare();
-
-  @override
-  _MiddleWareState createState() => _MiddleWareState();
-}
-
-class _MiddleWareState extends State<MiddleWare> {
-  @override
-  void initState() {
-    super.initState();
-    PrintUtils.printGreen("tes green");
-    PrintUtils.printMagenta("tes magenta");
-    PrintUtils.printWarning("tes yellow");
-    PrintUtils.printWhite("tes white");
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      backgroundColor: Theme.of(context).colorScheme.whiteTheme,
-      appBar: AppBar(
-        title: const Text('AppBar Demo'),
-        actions: <Widget>[],
-      ),
-    );
-  }
-}
