+++ lib/pages/404/404_page.dart
@@ -1,0 +1,12 @@
+import 'package:flutter/material.dart';
+
+class NotFoundPage extends StatelessWidget {
+  const NotFoundPage({Key? key}) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    return Center(
+      child: Text('Not Found Page'),
+    );
+  }
+}
